MVP-4710-patch-7: clean-up and refactor NotifiTargetContext and target utils
