+++ packages/notifi-react/lib/utils/target.ts
@@ -11,6 +11,27 @@ import {
   ToggleTarget,
 } from '../context';
 
+export const formatTelegramForSubscription = (telegramId: string) => {
+  if (telegramId.startsWith('@')) {
+    return telegramId.slice(1);
+  }
+  return telegramId;
+};
+
+export const reformatSignatureForWalletTarget = (
+  signature: Uint8Array | string,
+) => {
+  if (!signature) return '';
+
+  let hexString = '0x';
+
+  Object.values(signature).forEach(
+    (v) => (hexString += v.toString(16).padStart(2, '0')),
+  );
+
+  return hexString;
+};
+
 export const hasTarget = (targetData: TargetData) => {
   return !objectKeys(targetData).every((key) => {
     const target = targetData[key];
